backend/requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
fsspec==2024.6.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.0.1
pillow==10.4.0
requests==2.32.3
setuptools==72.1.0
sympy==1.13.1
torch==2.4.0
torchvision==0.19.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
google-cloud-vision==2.7.1
wikipedia-api==0.5.4

